+++ src/components/aside/aside.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useNavigate, useLocation } from 'react-router';
 
-import UserProfile from './user-profile';
+import UserProfile from '../user-profile';
 import { Asidetab } from './aside-tab-menu';
 
 import { MessageCircleMore } from 'lucide-react';
